[cudacompat] Run everything on CPU via cudacompat

+++ src/cudacompat/plugin-SiPixelClusterizer/gpuCalibPixel.h
@@ -35,7 +35,7 @@ namespace gpuCalibPixel {
     // zero for next kernels...
     if (0 == first)
       clusModuleStart[0] = moduleStart[0] = 0;
-    for (int i = first; i < gpuClustering::MaxNumModules; i += gridDim.x * blockDim.x) {
+    for (int i = first; i < static_cast<int>(gpuClustering::MaxNumModules); i += gridDim.x * blockDim.x) {
       nClustersInModule[i] = 0;
     }
